# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMS manager",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/cms-template/public",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Display the specified order by ID",
                "operationId": "e7e6d283779bcb33a966e6b60dc73b87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update the specified order",
                "operationId": "6a871b138ad8f7767d5b5031385ce402",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to update order"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete the specified order",
                "operationId": "65fa20f05c86dcf47c4aa4d038aa991b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted"
                    },
                    "400": {
                        "description": "Failed to delete order"
                    }
                }
            }
        },
        "/api/admin/orders/order-number/{orderNumber}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Display the specified order by order number",
                "operationId": "95cc7bd47f4c1fa88005f96d602ef837",
                "parameters": [
                    {
                        "name": "orderNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "ac9de7f6338d9523f477427635b5280e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/orders/customer/{customerId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders for a specific customer",
                "operationId": "5f86b7a80b6e3aaf7f022ea93273445b",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer's orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/orders/recent/{limit}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get the most recent orders",
                "operationId": "275e8b40410a99099012fb962c7ec0b8",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recent orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/orders/{id}/paid": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mark an order as paid",
                "operationId": "0d6a05ab409210650f7514036895f185",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order marked as paid"
                    },
                    "400": {
                        "description": "Failed to mark order as paid"
                    }
                }
            }
        },
        "/api/admin/orders/{orderId}/items": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create order items from cart items",
                "operationId": "56d68b4065c2daa68591ad1bcc3a81fd",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/OrderItem"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order items created"
                    }
                }
            }
        },
        "/api/admin/orders/{orderId}/items/{itemId}": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order item",
                "operationId": "4e6f1656058af80a2ce40f803863c578",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order item deleted"
                    },
                    "404": {
                        "description": "Order item not found"
                    }
                }
            }
        },
        "/api/admin/orders/{id}/status": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update the status of an order",
                "operationId": "9ce4bdb28ea07a751a86d4f5ca1e0840",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order status updated"
                    },
                    "400": {
                        "description": "Failed to update order status"
                    }
                }
            }
        },
        "/api/admin/orders/{id}/payment-status": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update the payment status of an order",
                "operationId": "387d47cdfc8ce24477e56541a7c3d742",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment status updated"
                    },
                    "400": {
                        "description": "Failed to update payment status"
                    }
                }
            }
        },
        "/api/admin/orders/{id}/payment-method": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update the payment method of an order",
                "operationId": "a346f5d2cfc33ae5a9323fb0256595cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment method updated"
                    },
                    "400": {
                        "description": "Failed to update payment method"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order",
                "description": "Order model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-20250417-001"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "payment_status": {
                        "type": "string",
                        "example": "unpaid"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "card"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "tax": {
                        "type": "number",
                        "format": "float",
                        "example": 12
                    },
                    "shipping_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 5
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 157
                    },
                    "notes": {
                        "type": "string",
                        "example": "Leave at the front door"
                    },
                    "placed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-17T14:00:00Z"
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-17T14:15:00Z"
                    },
                    "is_expired": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Authentication using Bearer token (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieSession": {
                "type": "apiKey",
                "description": "Authentication using session cookie (laravel_session)",
                "name": "laravel_session",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}